+++ lib/src/models/structured_document_tag_update.dart
@@ -69,8 +69,8 @@ class StructuredDocumentTagUpdate extends StructuredDocumentTag {
 
     if (json.containsKey('Appearance')) {
       switch (json['Appearance'] as String) {
-        case 'Default': appearance = StructuredDocumentTagBase_AppearanceEnum.defaultValue; break;
         case 'BoundingBox': appearance = StructuredDocumentTagBase_AppearanceEnum.boundingBox; break;
+        case 'Default': appearance = StructuredDocumentTagBase_AppearanceEnum.defaultValue; break;
         case 'Tags': appearance = StructuredDocumentTagBase_AppearanceEnum.tags; break;
         case 'Hidden': appearance = StructuredDocumentTagBase_AppearanceEnum.hidden; break;
         default: appearance = null; break;